--- Data/mongod.lock
@@ -1,1 +1,0 @@
-1
