--- app/index_2.php
@@ -1,48 +1,0 @@
-<!DOCTYPE html>
-<html lang="fr">
-  <head>
-    <!-- Required meta tags -->
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-
-    <!-- Bootstrap CSS -->
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0-beta/css/bootstrap.min.css" integrity="sha384-/Y6pD6FV/Vv2HJnA6t+vslU6fwYXjCFtcEpHbNJ0lyAFsXTsjBbfaDjzALeQsN6M" crossorigin="anonymous">
-  </head>
-  <body>
-      <header>
-    <div class="container">
-      <div class="row">
-        <div class="col">
-          <h1>CONNECTION</h1>
-        </div>
-      </div>
-      <hr>
-      </div>
-          </header>
-
-      <div class="container">
-          <div class="row">
-              <div class="col-3"></div>
-
-
-        <div class="col-6" style="align:center">
-         <form method="post" action="index2.php">
-          <div class="form-group">
-            <label for="Email">Email</label>
-            <input type="Email" class="form-control" name="mail" id="mail" placeholder="mail">
-          </div>
-          <div class="form-group">
-            <label for="password">Password</label>
-            <input type="text" class="form-control" name="password" id="password" placeholder="text">
-          </div>
-
-             <div class="row justify-content-md-center">
-          <button class="btn btn-primary">Submit</button>
-             </div>
-        </form>
-       </div>
-              <div class="col-3"></div>
-          </div>
-      </div>
-  </body>
-</html>
