--- docker-compose2.yml
@@ -1,22 +1,0 @@
-version: '3'
-
-services:
-  node:
-    build: ./
-    container_name: node
-    volumes:
-      - ./:/home/node/app
-    ports:
-      - 80:80
-    depends_on:
-      - mysql
-
-  mysql:
-    build: ./mysql
-    container_name: mysql
-    ports:
-      - 3306:3306
-    volumes:
-      - ./mysql/data:/var/lib/mysql
-
-
