+++ mysql/data/binlog.index
@@ -20,3 +20,6 @@
 ./binlog.000020
 ./binlog.000021
 ./binlog.000022
+./binlog.000023
+./binlog.000024
+./binlog.000025
