+++ mysql/data/ib_buffer_pool
@@ -1,4 +1,4 @@
-4294967279,281
+4294967279,273
 4294967293,131
 4294967293,130
 4294967293,129
@@ -140,6 +140,7 @@
 4294566154,0
 4294566153,0
 4294967293,0
+4294967278,234
 4294967278,109
 4294967278,233
 4294967278,232
@@ -234,5 +235,3 @@
 4294967278,69
 4294967278,181
 4294967278,68
-4294967278,180
-4294967278,67
